moved NoResults/ into components/
