+++ site/src/component/NoResults/NoResults.scss
